完成视频详情页
内容详情前端项目

+++ gec-auth-parent/gec-model/src/main/java/com/lanxige/Dto/VideoDanmakuDTO.java
@@ -22,7 +22,7 @@ public class VideoDanmakuDTO extends BaseEntity {
     /**
      * 出现时间(秒)
      */
-    private Integer playTime;
+    private Double playTime;
 
     /**
      * 颜色
